+++ game/logic/changshu/majiang/hulib.go
@@ -19,7 +19,7 @@ func init() {
 type HuTypeList []EmHuType
 
 func (self *HuLib) normal_hu(cardInfo *PlayerCardInfo) (bool, EmHuType) {
-	var var2key = map[byte]int{
+	var var2key = map[int32]int{
 		11: 1, 12: 2, 13: 3, 14: 4, 15: 5, 16: 6, 17: 7, 18: 8, 19: 9,
 		21: 10, 22: 11, 23: 12, 24: 13, 25: 14, 26: 15, 27: 16, 28: 17, 29: 18,
 		31: 19, 32: 20, 33: 21, 34: 22, 35: 23, 36: 24, 37: 25, 38: 26, 39: 27,
@@ -28,7 +28,7 @@ func (self *HuLib) normal_hu(cardInfo *PlayerCardInfo) (bool, EmHuType) {
 	checkcard := make([]int, 34)
 	for card, num := range cardInfo.StackCards {
 		if key, ok := var2key[card]; ok {
-			checkcard[key] = num
+			checkcard[key] = int(num)
 		} else {
 			log.Errorf("牌值%v不存在,请检查!!!", card)
 		}
